+++ app/Http/Resources/PickupLocation/PickupLocationResource.php
@@ -20,6 +20,7 @@ class PickupLocationResource extends JsonResource
             'working_hours' => $this->working_hours,
             'status' => (bool) $this->status,
             'display_order' => $this->display_order,
+            'is_default' => (bool) $this->is_default,
         ];
     }
 }